requirements.txt
python-telegram-bot==21.6
groq==0.11.0
mcp==1.1.2
httpx==0.27.2
